Implicit cursors
Perspective with a query to two related tables (This will be used to validate the entered data). A PL/SQL procedure has been added to allow simultaneous data entry into two tables. The procedure uses the PRAGMA directive to isolate transactions and check the state of the primary key sequence in the source table. A PL/SQL procedure was added to allow concurrent data entry into tables based on the previously added procedure, but using an implicit cursor (RETURNING) and IF SQL% cursor validation. The operation of the procedure was checked by calling it and querying the corresponding perspective.